src/api-gateway/pyproject.toml
[project]
name = "api-gateway"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.127.0",
    "httpx>=0.28.1",
    "pyyaml>=6.0.3",
]
